Strikeクラス, Spareクラスを削除、ストライクとスペアの各合計をクラスメソッドを使ってFrameクラスから呼び出すように修正
