+++ 07.bowling_object/lib/frame.rb
@@ -11,4 +11,24 @@ class Frame
     @next_frame = frames[index + 1]
     @after_next_frame = frames[index + 2]
   end
+
+  def self.calc_total_strike(frames)
+    strikes = []
+    frames.first(9).each do |fr|
+      if fr.first_shot.mark == 10 && fr.next_frame[0] == 10
+        strikes << [fr.next_frame[0], fr.after_next_frame[0]].sum
+      elsif fr.first_shot.mark == 10
+        strikes << fr.next_frame.sum
+      end
+    end
+    strikes.sum
+  end
+
+  def self.calc_total_spare(frames)
+    spares = []
+    frames.first(9).each do |fr|
+      spares << fr.next_frame[0] if [fr.first_shot.mark, fr.second_shot.mark].compact.sum == 10 && !fr.second_shot.mark.nil?
+    end
+    spares.sum
+  end
 end
