+++ 07.bowling_object/test/frame_test.rb
@@ -12,4 +12,20 @@ class FrameTest < Minitest::Test
 
     assert_equal frames[0].after_next_frame, frames[1].next_frame
   end
+
+  def test_calc_total_strike
+    ARGV.replace(['6,3,9,0,0,3,8,2,7,3,X,9,1,8,0,X,6,4,5'])
+    game = Game.new
+    frames = game.create_frames_with_new
+
+    assert_equal 20, game.calc_total_strike(frames)
+  end
+
+  def test_calc_total_spare
+    ARGV.replace(['6,3,9,0,0,3,8,2,7,3,X,9,1,8,0,X,6,4,5'])
+    game = Game.new
+    frames = game.create_frames_with_new
+
+    assert_equal 25, game.calc_total_spare(frames)
+  end
 end
