+++ 07.bowling_object/test/game_test.rb
@@ -26,8 +26,8 @@ class GameTest < Minitest::Test
   def test_calc_total_score
     shots = @game.convert_scores_to_shots
     frames = @game.create_frames_with_new
-    total_strike = Strike.new(frames).calc_total_strike
-    total_spare = Spare.new(frames).calc_total_spare
+    total_strike = @game.calc_total_strike(frames)
+    total_spare = @game.calc_total_spare(frames)
 
     assert_equal 139, @game.calc_total_score(shots, total_strike, total_spare)
   end
